requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
astroid==2.5.1
autopep8==1.5.5
bcrypt==3.2.0
blis==0.7.4
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.0.0/en_core_web_trf-3.0.0-py3-none-any.whl
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-HTTPAuth==4.2.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
ftfy==5.9
gunicorn==20.1.0
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
lazy-object-proxy==1.5.2
lxml==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
murmurhash==1.0.5
numpy==1.20.1
packaging==20.9
pathy==0.4.0
preshed==3.0.5
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
pyflakes==2.2.0
pylint==2.7.2
pyparsing==2.4.7
PyPDF2==1.26.0
python-docx==0.8.10
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
six==1.15.0
smart-open==3.0.0
spacy==3.0.5
spacy-alignments==0.7.2
spacy-legacy==3.0.1
spacy-transformers==1.0.1
SQLAlchemy==1.3.23
srsly==2.4.0
thinc==8.0.2
tokenizers==0.9.4
toml==0.10.2
torch==1.8.1+cpu
torchcontrib==0.0.2
tqdm==4.59.0
transformers==4.2.2
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
